requirements.txt
streamlit
chromadb

langchain==0.1.17
langchain-community==0.0.36
langchain-google-genai==1.0.4

sentence-transformers

python-dotenv
pymupdf
tqdm